[Auth/Core] Character Create & Delete finished Characters are saved to the DB Characters can be marked as deleted SelectReply packet updated with Tint info AttachmentSlots update for v0.179.1.71385 (Current Version of Client used) Database migrations to add skin tint, and fixed up bodytype For fun, generate a random chest piece to display in character select
Added comment for clarity of logic

+++ LandmarkEmulator.Database.Character/CharacterDatabase.cs
@@ -84,7 +84,7 @@ namespace LandmarkEmulator.Database.Character
             return await context.Character.FirstOrDefaultAsync(e => e.Name == name);
         }
 
-        public async Task<List<CharacterModel>> GetCharacters(uint accountId)
+        public async Task<List<CharacterModel>> GetCharacters(ulong accountId)
         {
             using var context = new CharacterContext(config);
             return await context.Character.Where(c => c.AccountId == accountId)
